requirements.txt
pillow>=9.0.0
pyinstaller>=5.0.0 